Changes that simplify running the tests under Bazel.
Changed all paths to be absolute, removed `os.chdir` call.  For some
reason `chdir` was not working for me -- I suspect this might be related
to the fact the "current working directory" is a process-global concept,
and it's hard to guarantee that it would not be changed by something in
other thread, or even by the same thread, in the time between module
load and actual test execution.

The way I'm trying to run the tests, it's reading the test data from
read-only mounted partition, so it's not possible to write "*.got" files
near the "*.right" files.  So I added a way
(`XDIS_DONT_WRITE_DOT_GOT_FILES` environment variable) to simple disable
writing those files.