Add switch to allow showing dup line numbers
Python's disassembly doesn't show line numbers other than the first one
when there is more entry for a given line. (That's probably a bug
that should be reported?)

Right now we show the full deal in pydisasm for asm format.
Here it is imporant since we recreate the line number table based
on information given in the instructions.

We could and probably should allow showing all of the line number in the
default format as well.

Underneath this commit adds the parameter to control that.