check_load(): better python source detection...
check_load() previously we relied on the extension being ".py" to detect if a file
was Python source. There are many valid Python programs that don't have
a ".py" extension.

With this change we compile the passed file and that determines if we
have Python source code.

Added/exported is_python_source() function